requirements.txt
 pandas
numpy
pyarrow
duckdb
psycopg2-binary
sqlalchemy